apps/form_bot/requirements.txt
discord.py>=2.6.0
python-dotenv>=1.1.0
supabase>=2.22.0
asyncpg>=0.30.0
python-dateutil>=2.9.0
httpx>=0.28.0
aiohttp>=3.13.0

